\name{costError}
\alias{costEntropy}
\alias{costError}
\alias{costGini}
\title{Cost Functions}
\usage{
  costError(y)

  costGini(y)

  costEntropy(y)
}
\arguments{
  \item{y}{a factor.}
}
\value{
  a numeric cost.
}
\description{
  These functions compute the cost of representing a set
  with its majority element.
}

