\name{crossValidate}
\alias{crossValidate}
\title{Cross-Validation Framework}
\usage{
  crossValidate(data, tuning, train, validate, folds = 10L)
}
\arguments{
  \item{data}{}

  \item{tuning}{}

  \item{train}{}

  \item{validate}{}

  \item{folds}{}
}
\description{
  This function performs k-fold cross-validation,
  optionally over the range of some tuning parameter.
}

