\name{print.ClassForest}
\alias{print.ClassForest}
\title{Print Method For Random Forests}
\usage{
  \method{print}{ClassForest} (x, ...)
}
\arguments{
  \item{x}{a ClassForest object, which will be printed to
  the console.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  Print Method For Random Forests
}

