\name{splitDetails}
\alias{splitDetails}
\title{Retrieve Split Details}
\usage{
  splitDetails(y)
}
\arguments{
  \item{y}{a factor, containing the true classes of the
  split observations.}
}
\value{
  a list, containing the name of the majority class and the
  number of observations in each class.
}
\description{
  \code{splitDetails} retrieves the number of elements and
  the decision for a split.
}

